requirements.txt
tqdm
trimesh
smplx
tensorboardX
torchgeometry
pyquaternion
openai-clip
einops
wis3d

# for stage 1
# shapely
# tenacity
# openai
# scikit-learn